+++ _example/simple-echo/cjk-cyrillic/main.go
@@ -23,7 +23,7 @@ func completer(in prompt.Document) []prompt.Suggest {
 func main() {
 	p := prompt.New(
 		executor,
-		completer,
+		prompt.WithCompleter(completer),
 		prompt.WithPrefix(">>> "),
 		prompt.WithTitle("sql-prompt for multi width characters"),
 	)