+++ CHANGELOG.md
@@ -38,7 +38,10 @@ This release aims to make the code a bit cleaner, fix a couple of bugs and provi
 - The cursor can move in 2D.
 - The Up arrow key will jump to the line above if the cursor is beyond the first line, but it will replace the input with the previous history entry if it's on the first line (like in Ruby's irb)
 - The Down arrow key will jump to the line below if the cursor is before the last line, but it will replace the input with the next history entry if it's on the last line (like in Ruby's irb)
+- <kbd>Tab</kbd> will insert a single indentation level when there are no suggestions
+- <kbd>Shift</kbd> + <kbd>Tab</kbd> will delete a single indentation level when there are no suggestions and the line before the cursors consists only of indentation (spaces)
 - Make `Completer` optional when creating a new `prompt.Prompt`. Change the signature of `prompt.New` from `func New(Executor, Completer, ...Option) *Prompt` to `func New(Executor, ...Option) *Prompt`
+- Make `prefix` optional in `prompt.Input`. Change the signature of `prompt.Input` from `func Input(string, ...Option) string` to `func Input(...Option) string`.
 - Rename `prompt.ConsoleParser` to `prompt.Reader` and make it embed `io.ReadCloser`
 - Rename `prompt.ConsoleWriter` to `prompt.Writer` and make it embed `io.Writer` and `io.StringWriter`
 - Rename `prompt.OptionTitle` to `prompt.WithTitle`